movie_recommendation_system/requirements.txt
Jinja2==3.1.6
joblib==1.5.3
nltk==3.9.2
numpy==2.3.5
pandas==2.3.3
pandas-stubs==2.3.3.251219
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.8.0
streamlit==1.52.2
tzdata==2025.3

